+++ app/src/main/java/com/digitalsolution/familyfilmapp/ui/screens/login/uistates/LoginRegisterState.kt
@@ -1,0 +1,22 @@
+package com.digitalsolution.familyfilmapp.ui.screens.login.uistates
+
+import androidx.annotation.StringRes
+import com.digitalsolution.familyfilmapp.R
+
+sealed class LoginRegisterState(
+    @StringRes val buttonText: Int,
+    @StringRes val accountText: Int,
+    @StringRes val signText: Int
+) {
+    data class Login(
+        @StringRes val button: Int = R.string.login_text_button,
+        @StringRes val account: Int = R.string.login_text_no_account,
+        @StringRes val sign: Int = R.string.login_text_sign_up
+    ) : LoginRegisterState(button, account, sign)
+
+    data class Register(
+        @StringRes val button: Int = R.string.register_text_button,
+        @StringRes val account: Int = R.string.login_text_yes_account,
+        @StringRes val sign: Int = R.string.login_text_sign_in
+    ) : LoginRegisterState(button, account, sign)
+}
