--- app/src/main/res/values-night/strings.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="app_name">Family Film</string>
-
-    <!-- LOGIN-->
-    <string name="login_snail_logo">Snail Logo</string>
-    <string name="login_text_app_title">Film Family</string>
-    <string name="login_text_app_subtitle">Shared your films with your family and friends</string>
-    <string name="login_text_field_email">Enter your email</string>
-    <string name="login_text_field_password">Enter your password</string>
-    <string name="login_text_button">Login</string>
-    <string name="register_text_button">Register</string>
-    <string name="login_text_sign_in_with_google">Sign in with Google</string>
-    <string name="login_icon_google">Google Logo</string>
-    <string name="login_text_no_account">Don\'t have an account?</string>
-    <string name="login_text_yes_account">Do you have an account?</string>
-    <string name="login_text_sign_up">Sign up</string>
-    <string name="login_text_sign_in">Sign in</string>
-    <string name="login_text_forgot_your_password">Forgot your password?</string>
-    <string name="login_text_recover_password">Introduce the email for recover your password</string>
-    <string name="login_text_send_recover_password">Send</string>
-    <string name="login_text_recover_password_cancel">Cancel</string>
-
-    <!-- HOME-->
-    <string name="home_text_my_list">My List</string>
-    <string name="home_text_seen">Seen</string>
-    <string name="details_screen_description_lorem_ipsum">"\n                    Lorem ipsum dolor sit amet consectetur adipiscing elit, nunc nibh erat cum ultrices sociis, nisi facilisi per mollis torquent laoreet. Ultricies tellus cubilia vulputate iaculis varius, aliquam enim velit egestas. Malesuada etiam tempus lectus facilisis nam at ultricies, mus fermentum sollicitudin natoque sodales.\n\n                    Effendi facilisi integer magna porttitor rutrum sodales aliquet volutpat, porta aenean lectus nisi ridiculus penatibus leo rhoncus semper, sagittis libero dignissim pellentesque urna fames in. Natoque volutpat imperdiet dis quisque cum iaculis, senectus euismod molestie mollis cubilia purus, dictumst cras nec platea dictum. Mollis sodales mus sed ligula condimentum dui aliquet, sociis scelerisque senectus commodo convallis taciti, donec hac gravida cursus id viverra. \n\n                    Lorem ipsum dolor sit amet consectetur adipiscing elit, nunc nibh erat cum ultrices sociis, nisi facilisi per mollis torquent laoreet. Ultricies tellus cubilia vulputate iaculis varius, aliquam enim velit egestas. Malesuada etiam tempus lectus facilisis nam at ultricies, mus fermentum sollicitudin natoque sodales.\n\n                    Effendi facilisi integer magna porttitor rutrum sodales aliquet volutpat, porta aenean lectus nisi ridiculus penatibus leo rhoncus semper, sagittis libero dignissim pellentesque urna fames in. Natoque volutpat imperdiet dis quisque cum iaculis, senectus euismod molestie mollis cubilia purus, dictumst cras nec platea dictum. Mollis sodales mus sed ligula condimentum dui aliquet, sociis scelerisque senectus commodo convallis taciti, donec hac gravida cursus id viverra. \n\n                    Eleifend facilisi integer magna porttitor rutrum sodales aliquet volutpat, porta aenean lectus nisi ridiculus penatibus leo rhoncus semper, sagittis libero dignissim pellentesque urna fames in. Natoque volutpat imperdiet dis quisque cum iaculis, senectus euismod molestie mollis cubilia purus, dictumst cras nec platea dictum. Mollis sodales mus sed ligula condimentum dui aliquet, sociis scelerisque senectus commodo convallis taciti, donec hac gravida cursus id viverra.\n                "</string>
-
-</resources>
